Refactoring the namespaces a litte bit
