+++ ALifeUniv/ALife/Scenarios/Mazes/MazeScenario.cs
@@ -11,6 +11,7 @@
  * **/
 
 using ALifeUni.ALife.Brains;
+using ALifeUni.ALife.Scenarios.ScenarioHelpers;
 using ALifeUni.ALife.Shapes;
 using ALifeUni.ALife.Utility;
 using System;
@@ -165,7 +166,7 @@ namespace ALifeUni.ALife.Scenarios
             }
             //MazeRunner mr = new MazeRunner(red, blue);
 
-            MazeScenarioHelpers.SetUpMaze();
+            MazeSetups.SetUpMaze();
         }
 
         int bestXNum = 5;
