+++ ALifeUniv/UI/UserControls/NeuralBrainView.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl
-    x:Class="ALifeUni.NeuralBrainView"
+    x:Class="ALifeUni.UI.UserControls.NeuralBrainView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:canvas="using:Microsoft.Graphics.Canvas.UI.Xaml"
