+++ ALifeUniv/UI/UserControls/NeuralBrainView.xaml.cs
@@ -21,7 +21,7 @@ using Windows.UI.Xaml.Shapes;
 
 // The User Control item template is documented at https://go.microsoft.com/fwlink/?LinkId=234236
 
-namespace ALifeUni
+namespace ALifeUni.UI.UserControls
 {
     public sealed partial class NeuralBrainView : UserControl
     {
